docs: Refactoring away from GlobalHeader example
Focus more on the live badges of the icon buttons. GlobalHeader is just a distraction.
