+++ modules/react/_examples/stories/AriaLiveRegions.stories.mdx
@@ -2,7 +2,7 @@ import {AriaLiveRegion} from '@workday/canvas-kit-react/common';
 import {FilterListWithLiveStatus} from './examples/common/FilterListWithLiveStatus';
 import {VisibleLiveRegion} from './examples/common/VisibleLiveRegion';
 import {HiddenLiveRegion} from './examples/common/HiddenLiveRegion';
-import {GlobalHeaderWithLiveBadge} from './examples/common/GlobalHeaderWithLiveBadge';
+import {IconButtonsWithLiveBadges} from './examples/common/IconButtonsWithLiveBadges';
 
 <Meta title="Examples/AriaLiveRegion" component={AriaLiveRegion} />
 
@@ -22,4 +22,4 @@ import {GlobalHeaderWithLiveBadge} from './examples/common/GlobalHeaderWithLiveB
 
 ## Global header with live badge
 
-<ExampleCodeBlock code={GlobalHeaderWithLiveBadge} />
+<ExampleCodeBlock code={IconButtonsWithLiveBadges} />
