+++ modules/react/_examples/stories/examples/common/IconButtonsWithLiveBadges.tsx
@@ -1,0 +1,84 @@
+import React, {useState} from 'react';
+import {AccessibleHide, AriaLiveRegion, useUniqueId} from '@workday/canvas-kit-react/common';
+import {
+  notificationsIcon,
+  inboxIcon,
+  assistantIcon,
+  userIcon,
+} from '@workday/canvas-system-icons-web';
+import {space} from '@workday/canvas-kit-react/tokens';
+import {SecondaryButton, TertiaryButton} from '@workday/canvas-kit-react/button';
+import {Flex} from '@workday/canvas-kit-react/layout';
+import {Tooltip} from '@workday/canvas-kit-react/tooltip';
+import {CountBadge} from '@workday/canvas-kit-react/badge';
+
+const MyTasksLiveBadge = ({cnt}) => {
+  // use tooltip to assign name,
+  // use AriaLiveRegion inside button,
+  // assign name to live region referencing the button,
+  // use BadgeCount inside live region,
+  // use AccessibleHide to create invisible word "new" after badge
+  // use aria-describedby on button, referencing live region container to set description
+  // :( isn't working at all in Safari
+  // :( isn't working well in Firefox
+  const badgeID = useUniqueId();
+  const myTasksID = useUniqueId();
+
+  return (
+    <Tooltip title="My Tasks">
+      <TertiaryButton icon={inboxIcon} id={myTasksID} aria-describedby={badgeID}>
+        <AriaLiveRegion id={badgeID} aria-labelledby={myTasksID}>
+          <CountBadge count={cnt} />
+          <AccessibleHide>New</AccessibleHide>
+        </AriaLiveRegion>
+      </TertiaryButton>
+    </Tooltip>
+  );
+};
+
+// use AriaLiveRegion around the button,
+// use Tooltip to assign the name of the button,
+// make sure Tooltip title string includes count value
+// Chrome + VO => Announces name "notifications X new" and innerText 'X'
+// Safari + VO => Works nicely :) announces button name
+const NotificationsLiveBadge = ({cnt}) => (
+  <AriaLiveRegion>
+    <Tooltip title={`Notifications ${cnt} new`}>
+      <TertiaryButton icon={notificationsIcon}>
+        <CountBadge count={cnt} />
+      </TertiaryButton>
+    </Tooltip>
+  </AriaLiveRegion>
+);
+
+export const IconButtonsWithLiveBadges = () => {
+  const [counter, setCounter] = useState(0);
+  const [notifications, setNotifications] = useState(0);
+
+  const handleAddTask = () => {
+    setCounter(prev => prev + 1);
+  };
+
+  const handleAddNotification = () => {
+    setNotifications(prev => prev + 1);
+  };
+
+  return (
+    <>
+      <Flex as="header" gap={space.s}>
+        <Tooltip title="Workday Assistant">
+          <TertiaryButton icon={assistantIcon} />
+        </Tooltip>
+        <NotificationsLiveBadge cnt={notifications} />
+        <MyTasksLiveBadge cnt={counter} />
+        <Tooltip title="Profile">
+          <TertiaryButton icon={userIcon} />
+        </Tooltip>
+      </Flex>
+      <Flex padding={space.s} gap={space.s} as="main">
+        <SecondaryButton onClick={handleAddNotification}>Add a Notification</SecondaryButton>
+        <SecondaryButton onClick={handleAddTask}>Add an item to My Tasks</SecondaryButton>
+      </Flex>
+    </>
+  );
+};
